Refactor testing code.
Consolidates all testing functions into `test_all.py` which serves
both as an entry point for local testing and contains all of the
functions tested by py.test.

Begins to address #170 (work in progress)

Resolves: #157
